requirements.txt
asgiref==3.10.0
dj-database-url==3.0.1
Django==5.2.5
django-cors-headers==4.8.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
wheel==0.45.1
faker==18.9.1
stripe==11.12.0
pandas==2.3.3
scikit-learn==1.7.2
joblib==1.5.2
numpy==2.3.4
scipy==1.16.3
python-dateutil==2.9.0.post0
six==1.17.0
threadpoolctl==3.6.0

